fix StatisticNames

+++ src/main/java/com/melnikov/service/dto/StatisticNames.java
@@ -12,5 +12,5 @@ import lombok.Setter;
 @AllArgsConstructor
 @NoArgsConstructor
 public class StatisticNames {
-    private Integer usedNamesPercent;
+    private Double usedNamesPercent;
 }
